Prohibimos el acceso a contactos que no te pertenecen

+++ delete.php
@@ -20,6 +20,13 @@ if ($statement->rowCount() == 0) {
   return;
 }
 
+$contact = $statement->fetch(PDO::FETCH_ASSOC);
+if ($contact["user_id"] !== $_SESSION["user"]["id"]) {
+  http_response_code(403);
+  echo ("HTTP 403 UNAUTHORIZED");
+  return;
+}
+
 $conn->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
 
 header("Location: home.php");
